Refine How to Avoid Python Updates in Dockerfile
The previous logic tries to automerge by checking if a change is to a Python minor version. However, this logic causes automerge not to merge updates to python-decouple, which has the string "python" in it. The new logic takes advantage of a recent update to the fetch-metadata action that allows us to see if a change affects the "docker" or "pip" ecosystems, independently.